Refactor/129 arrange file and code (#130)
* refactor: 폴더 성격에 맞게 파일 이동

* refactor: 사용하지 않는 전역 상태 삭제

* refactor: Role 타입 interface 폴더로 이동
