+++ src/components/TeamInfoModal.tsx
@@ -2,7 +2,7 @@ import { Clock, FileText, Square, Users, X } from "lucide-react";
 import { Button } from "./ui/button";
 import { Dialog, DialogContent } from "./ui/dialog";
 import { Team } from "@/interface/teams";
-import { addImagePrefix } from "@/utils/Image/imagePrefix";
+import { addImagePrefix } from "@/utils/imagePrefix";
 
 interface TeamInfoModalProps {
   selectedTeam: Team | null;
