+++ src/interface/role.ts
@@ -1,0 +1,1 @@
+export type Role = "MEMBER" | "USER" | "ADMIN" | "NONUSER";
