+++ src/store/atom.ts
@@ -1,16 +1,11 @@
 import { atom } from "recoil";
 import { recoilPersist } from "recoil-persist";
 import { JwtHIStudyPayload } from "@/components/GoogleButton";
+import { Role } from "@/interface/role";
 
 const { persistAtom } = recoilPersist({
   key: "localStorage", //원하는 key 값 입력
   storage: localStorage,
-});
-
-export const darkState = atom<boolean>({
-  key: "darkMode",
-  default: false,
-  effects_UNSTABLE: [persistAtom],
 });
 
 export const isLoginState = atom<boolean>({
@@ -29,42 +24,7 @@ export const userLoginInfo = atom<JwtHIStudyPayload | null>({
   default: null,
 });
 
-export const isCodeModalState = atom<boolean>({
-  key: "codeModal",
-  default: false,
-});
-
-export const isDelete = atom<boolean>({
-  key: "deleteSnackbar",
-  default: false,
-  effects_UNSTABLE: [persistAtom],
-});
-
-export const selectState = atom<number>({
-  key: "sidebarSelect",
-  default: 1,
-  effects_UNSTABLE: [persistAtom],
-});
-
-export const isLoadingState = atom<boolean>({
-  key: "isLoading",
-  default: false,
-});
-
-//TODO: 타입 정의 필요
-export const groupAutoCompleteState = atom<any[]>({
-  key: "groupAutoComplete",
-  default: [],
-});
-
-export type Role = "MEMBER" | "USER" | "ADMIN" | "NONUSER";
-
 export const roleState = atom<Role>({
   key: "role",
   default: "NONUSER",
-});
-
-export const isShowFullImageState = atom<boolean>({
-  key: "isShowFullImage",
-  default: false,
 });
