+++ src/utils/compressImageFile.ts
@@ -1,6 +1,6 @@
 import imageCompression from "browser-image-compression";
 
-const compressedFile = async (file: File) => {
+const compressedImageFile = async (file: File) => {
   const options = {
     maxSizeMB: 1,
     maxWidthOrHeight: 1920,
@@ -21,4 +21,4 @@ const compressedFile = async (file: File) => {
   // }
 };
 
-export default compressedFile;
+export default compressedImageFile;
